Representation of chromatographic MS data

The Chromatogram class is designed to store chromatographic MS data, i.e. pairs of retention time and intensity values. Instances of the class can be created with the Chromatogram constructor function but in most cases the dedicated methods for OnDiskMSnExp and MSnExp objects extracting chromatograms should be used instead (i.e. the chromatogram method).

Chromatogram: create an instance of the Chromatogram class.

aggregationFun,aggregationFun<- get or set the aggregation function.

rtime returns the retention times for the rentention time - intensity pairs stored in the chromatogram.

intensity returns the intensity for the rentention time - intensity pairs stored in the chromatogram.

mz get the mz (range) of the chromatogram. The function returns a numeric(2) with the lower and upper mz value.

precursorMz get the mz of the precursor ion. The function returns a numeric(2) with the lower and upper mz value.

fromFile returns the value from the fromFile slot.

length returns the length (number of retention time - intensity pairs) of the chromatogram.

as.data.frame returns the rtime and intensity values from the object as data.frame.

filterRt: filters the chromatogram based on the provided retention time range.

clean: Removes unused 0-intensity data points. See clean documentation for more details and examples.

plot: plots a Chromatogram object.

msLevel returns the MS level of the chromatogram.

isEmpty returns TRUE for empty chromatogram or chromatograms with all intensities being NA.

productMz get the mz of the product chromatogram/ion. The function returns a numeric(2) with the lower and upper mz value.

Chromatogram(rtime = numeric(), intensity = numeric(), mz = c(NA_real_,
  NA_real_), filterMz = c(NA_real_, NA_real_), precursorMz = c(NA_real_,
  NA_real_), productMz = c(NA_real_, NA_real_), fromFile = integer(),
  aggregationFun = character(), msLevel = 1L)

aggregationFun(object)

# S4 method for Chromatogram
show(object)

# S4 method for Chromatogram
rtime(object)

# S4 method for Chromatogram
intensity(object)

# S4 method for Chromatogram
mz(object, filter = FALSE)

# S4 method for Chromatogram
precursorMz(object)

# S4 method for Chromatogram
fromFile(object)

# S4 method for Chromatogram
length(x)

# S4 method for Chromatogram
as.data.frame(x)

# S4 method for Chromatogram
filterRt(object, rt)

# S4 method for Chromatogram
clean(object, all = FALSE, na.rm = FALSE)

# S4 method for Chromatogram,ANY
plot(x, col = "#00000060", lty = 1,
  type = "l", xlab = "retention time", ylab = "intensity", main = NULL,
  ...)

# S4 method for Chromatogram
msLevel(object)

# S4 method for Chromatogram
isEmpty(x)

# S4 method for Chromatogram
productMz(object)

Arguments

rtime
numeric with the retention times (length has to be equal to the length of intensity).

intensity
numeric with the intensity values (length has to be equal to the length of rtime).

mz
numeric(2) representing the mz value range (min, max) on which the chromatogram was created. This is supposed to contain the real range of mz values in contrast to the filterMz below. If not applicable use mzrange = c(0, 0).

filterMz
numeric(2) representing the mz value range (min, max) that was used to filter the original object on mz dimension. If not applicable use filterMz = c(0, 0).

precursorMz
numeric(2) for SRM/MRM transitions. Represents the mz of the precursor ion. See details for more information.

productMz
numeric(2) for SRM/MRM transitions. Represents the mz of the product. See details for more information.

fromFile
integer(1) the index of the file within the OnDiskMSnExp or MSnExp from which the chromatogram was extracted.

aggregationFun
character string specifying the function that was used to aggregate intensity values for the same retention time across the mz range. Supported are "sum" (total ion chromatogram), "max" (base peak chromatogram), "min" and "mean".

msLevel
integer with the MS level from which the chromatogram was extracted.

object
A Chromatogram object.

filter
For mz: whether the mz range used to filter the original object should be returned (filter = TRUE), or the mz range calculated on the real data (filter = FALSE).

x
For as.data.frame and length: a Chromatogram object.

rt
For filterRt: numeric(2) defining the lower and upper retention time for the filtering.

all
For clean: logical(1) whether all 0 intensities should be removed (default is FALSE). See clean for more details and examples.

na.rm
For clean: logical(1) whether all NA intensities should be removed before cleaning the Chromatogram. Defaults to FALSE. See clean for more details and examples.

col
For plot: the color to be used for plotting.

lty
For plot: the line type. See plot for more details.

type
For plot: the type of plot. See plot for more details.

xlab
For plot: the x-axis label.

ylab
For plot: the y-axis label.

main
For plot: the plot title. If not provided the mz range will be used as plot title.

...
For plot: additional arguments to be passed to the plot function.

Details

The mz, filterMz, precursorMz and productMz are stored as a numeric(2) representing a range even if the chromatogram was generated for only a single ion (i.e. a single mz value). Using ranges for mz values allow this class to be used also for e.g. total ion chromatograms or base peak chromatograms.

The slots precursorMz and productMz allow to represent SRM (single reaction monitoring) and MRM (multiple SRM) chromatograms. As example, a Chromatogram for a SRM transition 273 -> 153 will have a @precursorMz = c(273, 273) and a @productMz = c(153, 153).

Slots

.__classVersion__,rtime,intensity,mz,filterMz,precursorMz,productMz,fromFile,aggregationFun,msLevel

See corresponding parameter above.

See also

Examples

## Create a simple Chromatogram object.
ints <- abs(rnorm(100, sd = 100))
rts <- seq_len(length(ints))
chr <- Chromatogram(rtime = rts, intensity = ints)
chr

#> Object of class: Chromatogram
#> length of object: 100
#> from file: 
#> mz range: [NA, NA]
#> rt range: [1, 100]
#> MS level: 1

## Extract intensities
intensity(chr)

#>   [1] 140.0043517  25.5317055 243.7263611   0.5571287  62.1552721 114.8411606
#>   [7] 182.1817661  24.7325302  24.4199607  28.2705449  55.3699384  62.8982042
#>  [13] 206.5024895 163.0989402  51.2426950 186.3011492  52.2012515   5.2601910
#>  [19]  54.2996343  91.4074827  46.8154420  36.2951256 130.4543545  73.7776321
#>  [25] 188.8504929   9.7445104  93.5847354   1.5950311  82.6788954 151.2399651
#>  [31]  93.5363190  17.6488611  24.3685465 162.3548883  11.2038083  13.3997013
#>  [37] 191.0087468  27.9237242  31.3445978 106.7307879   7.0034850  63.9123324
#>  [43]   4.9964899  25.1483443  44.4797116 275.5417575   4.6531380  57.7709069
#>  [49]  11.8194874 191.1720491  86.2086482  24.3236740  20.6087195   1.9177592
#>  [55]   2.9560754  54.9827542 227.4114857 268.2557184  36.1221255  21.3355750
#>  [61] 107.4345882  66.5088249 111.3952419  24.5896412 117.7563309  97.5850616
#>  [67] 106.5057320  13.1670635  48.8628809 169.9450568 147.0736306  28.4150344
#>  [73] 133.7320413  23.6696283 131.8293384  52.3909788  60.6748047  10.9935672
#>  [79]  17.2181715   9.0327287 192.4343341 129.8392759  74.8791268  55.6224329
#>  [85]  54.8257264 111.0534893 261.2334333  15.5693776  43.3889790  38.1951112
#>  [91]  42.4187575 106.3101996 104.8712620   3.8102895  48.6148920 167.2882611
#>  [97]  35.4361164  94.6347886 131.6826356  29.6640025

## Extract retention times
rtime(chr)

#>   [1]   1   2   3   4   5   6   7   8   9  10  11  12  13  14  15  16  17  18
#>  [19]  19  20  21  22  23  24  25  26  27  28  29  30  31  32  33  34  35  36
#>  [37]  37  38  39  40  41  42  43  44  45  46  47  48  49  50  51  52  53  54
#>  [55]  55  56  57  58  59  60  61  62  63  64  65  66  67  68  69  70  71  72
#>  [73]  73  74  75  76  77  78  79  80  81  82  83  84  85  86  87  88  89  90
#>  [91]  91  92  93  94  95  96  97  98  99 100

## Extract the mz range - is NA for the present example
mz(chr)

#> [1] NA NA

## plot the Chromatogram
plot(chr)

#> Error in as.double(y): cannot coerce type 'S4' to vector of type 'double'

## Create a simple Chromatogram object based on random values.
chr <- Chromatogram(intensity = abs(rnorm(1000, mean = 2000, sd = 200)),
        rtime = sort(abs(rnorm(1000, mean = 10, sd = 5))))
chr

#> Object of class: Chromatogram
#> length of object: 1000
#> from file: 
#> mz range: [NA, NA]
#> rt range: [0.002530252, 23.75382]
#> MS level: 1

## Get the intensities
head(intensity(chr))

#> [1] 1658.884 1828.917 1971.020 1935.111 1965.487 1752.787

## Get the retention time
head(rtime(chr))

#> [1] 0.002530252 0.004804333 0.086750187 0.102224373 0.114592272 0.155135059

## What is the retention time range of the object?
range(rtime(chr))

#> [1]  0.002530252 23.753823765

## Filter the chromatogram to keep only values between 4 and 10 seconds
chr2 <- filterRt(chr, rt = c(4, 10))

range(rtime(chr2))

#> [1] 4.031794 9.975626