Class

Comparing feature names of two comparable MSnSet instances.

Objects from the Class

Objects can be created with compfnames. The method compares the feature names of two objects of class "MSnSet". It prints a summary matrix of common and unique feature names and invisibly returns a list of FeatComp instances.

The function will compute the common and unique features for all feature names of the two input objects (featureNames(x) and feautreNames(y)) as well as distinct subsets as defined in the fcol1 and fcol2 feautre variables.

Slots

name:

Object of class "character" defining the name of the compared features. By convention, "all" is used when all feature names are used; otherwise, the respective levels of the feature variables fcol1 and fcol2.

common:

Object of class "character" with the common feature names.

unique1:

Object of class "character" with the features unique to the first MSnSet (x in compfname).

unique2:

Object of class "character" with the features unique to the seconn MSnSet (y in compfname).

all:

Object of class "logical" defining if all features of only a subset were compared. One expects that name == "all" when all is TRUE.

Methods

Accessors names, common, unique1 and unique2 can be used to access the respective FeatComp slots.

compfnames

signature(x = "MSnSet", y = "MSnSet", fcol1 = "character", fcol2 = "character", simplify = "logical", verbose = "logical"): creates the FeatComp comparison object for instances x and y. The feature variables to be considered to details feature comparison can be defined by fcol1 (default is "markers" and fcol2 for x and y respectively). Setting either to NULL will only consider all feature names; in such case, of simplify is TRUE (default), an FeatComp object is returned instead of a list of length 1. The verbose logical controls if a summary table needs to be printed (default is TRUE).

compfnames

signature(x = "list", y = "missing", ...): when x is a list of MSnSet instances, compfnames is applied to all element pairs of x. Additional parameters fcol1, fcol2, simplify and verbose are passed to the pairwise comparison method.

show

signature(object = "FeatComp"): prints a summary of the object.

See also

averageMSnSet to compuate an average MSnSet.

Examples

library("pRolocdata")

#> Error: package ‘MSnbase’ required by ‘pRolocdata’ could not be found

data(tan2009r1)

#> Warning: data set ‘tan2009r1’ not found

data(tan2009r2)

#> Warning: data set ‘tan2009r2’ not found

x <- compfnames(tan2009r1, tan2009r2)

#> Error in compfnames(tan2009r1, tan2009r2): object 'tan2009r1' not found

x[[1]]

#> Error in eval(expr, envir, enclos): object 'x' not found

x[2:3]

#> Error in eval(expr, envir, enclos): object 'x' not found

head(common(x[[1]]))

#> Error in common(x[[1]]): object 'x' not found

data(tan2009r3)

#> Warning: data set ‘tan2009r3’ not found

tanl <- list(tan2009r1, tan2009r2, tan2009r3)

#> Error in eval(expr, envir, enclos): object 'tan2009r1' not found

xx <- compfnames(tanl, fcol1 = NULL)

#> Error in compfnames(tanl, fcol1 = NULL): object 'tanl' not found

length(xx)

#> Error in eval(expr, envir, enclos): object 'xx' not found

tail(xx)

#> Error in tail(xx): object 'xx' not found

all.equal(xx[[15]],
          compfnames(tan2009r2, tan2009r3, fcol1 = NULL))

#> Error in all.equal(xx[[15]], compfnames(tan2009r2, tan2009r3, fcol1 = NULL)): object 'xx' not found

str(sapply(xx, common))

#> Error in sapply(xx, common): object 'xx' not found