MSnbase: Base Functions and Classes for MS-based Proteomics

MSnbase provides classes, methods and functions for visualisation, manipulation and processing of mass spectrometry data.

Important class are "MSnExp" (raw data file), "MSnSet" (quantitation data) and "ReporterIons" (reporter ions for labelled proteomics).

Other classes are "Spectrum" and the subclasses "Spectrum1" (for MS spectra) and "Spectrum2" (for MSMS spectra), "MIAPE" (Minimum Information about Proteomics Experiments) and "MSnProcess" (for processing information). These should however not be of direct utility to users.

If you have questions, want to rebort a bug or share suggestions, please file an issue at https://github.com/lgatto/MSnbase/issues, contact me directly or ask a question on the Bioconductor support forum https://support.bioconductor.org/.

References

Laurent Gatto and Kathryn S. Lilley, MSnbase - an R/Bioconductor package for isobaric tagged mass spectrometry data visualization, processing and quantitation, Bioinformatics 28(2), 288-289 (2012).

Gatto L. and Lilley K.S., Towards reproducible MSMS data preprocessing, quality control and quantification. BSPR/EBI Proteomics Meeting, Hinxton, United Kingdom, 13-15 July 2010, http://dx.doi.org/10.1038/npre.2010.5010.1.

See also

Introductory information, use cases and details are available from the vignettes:

The demo vignette describe an use-case using a dummy data set provided with the package. It can be accessed with vignette("MSnbase-demo", package = "MSnbase").

The development vignette describes the classes implemented in MSnbase and can be accessed with vignette("MSnbase-development", package = "MSnbase").

Details about input/outcupt capabilities and formats can be found in vignette("MSnbase-io", package = "MSnbase").

Complete listing of available documentation with library(help = "MSnbase").