Class Containing Measured Variables and Their Meta-Data

An NAnnotatedDataFrame is an "AnnotatedDataFrame", as defined in the 'Biobase' package that includes additional labels for multiplexing annotation.

Objects from the Class

See "AnnotatedDataFrame" for object creation with new. Multiplexing data is defined by setting the multiplex and multiLables parameters.

Slots

multiplex:

Object of class "numeric" indicating the number of multiplexed samples described.

multiLabels:

Object of class "character" describing the multiplexing.

varMetadata:

Object of class "data.frame" with number of rows equal number of columns in data, and at least one column, named labelDescription, containing a textual description of each variable. Inherited from "AnnotatedDataFrame".

data:

Object of class "data.frame" containing samples (rows) and measured variables (columns). Inherited from "AnnotatedDataFrame".

dimLabels:

Object of class "character" of length 2 that provides labels for the rows and columns in the show method. Inherited from "AnnotatedDataFrame".

.__classVersion__:

Object of class "Versions" describing the instance version. Intended for developer use. Inherited from "AnnotatedDataFrame".

Extends

Class "AnnotatedDataFrame", directly. Class "Versioned", by class "AnnotatedDataFrame", distance 2.

Methods

dim

signature(object = "NAnnotatedDataFrame"): Returns the number of samples, variables and multiplex cardinality in the object.

multiplex

signature(object = "NAnnotatedDataFrame"): Returns the number of multipexed samples described by the object.

multiLabels

signature(object = "NAnnotatedDataFrame"): Returns the multiplex labels.

show

signature(object = "NAnnotatedDataFrame"): Textual description of the object.

See also

"AnnotatedDataFrame".

Examples

df <- data.frame(x=1:3,
                 y=LETTERS[1:3],
                 row.names=paste("Sample",1:3,sep=""))
metaData <-
  data.frame(labelDescription=c(
               "Numbers",
               "Factor levels"))
mplx <- c("M1","M2")
new("NAnnotatedDataFrame",
    data=df,
    varMetadata=metaData,
    multiplex=length(mplx),
    multiLabels=mplx)

#> An object of class 'NAnnotatedDataFrame'
#>   rowNames: Sample1 Sample2 Sample3
#>   varLabels: x y
#>   varMetadata: labelDescription
#>   Multiplexing: 2 - M1 M2