Adds Identification Data

This methods add identification data to an experiment "MSnExp" or to a "MSnSet".

Methods

signature(object = "MSnExp", id = "character", fcol = "character", icol = "character", verbose = "logical")

Adds the identification data stored in mzIdentML files to a "MSnExp" instance. The method handles one or multiple mzIdentML files provided via id. id has to be a character vector of valid filenames. fcol and icol specify the columns in the featureData slot and the identification data.frame that are used to merge both data together as character vectors (defaults are fcol = c("spectrum.file", "acquisition.number") and icol = c("spectrumFile", "acquisitionnum")). The verbose argument (default is TRUE) defines whether status messages should be showed.

signature(object = "MSnExp", id = "mzID", fcol = "character", icol = "character", verbose = "logical")

Same as above but id could be an mzID object.

signature(object = "MSnExp", id = "mzIDCollection", fcol = "character", icol = "character", verbose = "logical")

Same as above but id could be an mzIDCollection object.

signature(object = "MSnExp", id = "mzRident", fcol = "character", icol = "character", verbose = "logical")

Same as above but id could be a single mzRident object generated by mzR::openIdfile.

signature(object = "MSnExp", id = "data.frame", fcol = "character", icol = "character", verbose = "logical")

Same as above but id could be a data.frame.

signature(object = "MSnSet", id = "character", verbose = "logical")

Adds the identification data stored in mzIdentML files to an "MSnSet" instance. The method handles one or multiple mzIdentML files provided via id. id has to be a character vector of valid filenames. fcol and icol specify the columns in the featureData slot and the identification data.frame that are used to merge both data together as character vectors (defaults are fcol = c("spectrum.file", "acquisition.number") and icol = c("spectrumFile", "acquisitionnum")). The verbose argument (default is TRUE) defines whether status messages should be showed.

signature(object = "MSnSet", id = "mzID", fcol = "character", icol = "character", verbose = "logical")

Same as above but id could be an mzID object.

signature(object = "MSnSet", id = "mzIDCollection", fcol = "character", icol = "character", verbose = "logical")

Same as above but id could be an mzIDCollection object.

signature(object = "MSnSet", id = "data.frame", fcol = "character", icol = "character", verbose = "logical")

Same as above but id could be a data.frame.

Details

The featureData slots in an "MSnExp" or an "MSnSet" instance provides only one row per MS2 spectrum but the identification is not always bijective. If multiple possible matches are present only the highest ranked identification is added.

The column nprot contains the number of members in the protein group; the columns accession and description contain a semicolon separated list of all matches sorted by their rank values. The columns npsm.prot and npep.prot represent the number of PSMs and peptides that were matched to a particular protein group. The column npsm.pep indicates how many PSMs were attributed to a peptide (as defined by its sequence pepseq).

See also

MSnExp and MSnSet.

Examples

## find path to a mzXML file
quantFile <- dir(system.file(package = "MSnbase", dir = "extdata"),
                 full.name = TRUE, pattern = "mzXML$")
## find path to a mzIdentML file
identFile <- dir(system.file(package = "MSnbase", dir = "extdata"),
                 full.name = TRUE, pattern = "dummyiTRAQ.mzid")

## create basic MSnExp
msexp <- readMSData(quantFile)

## add identification information
msexp <- addIdentificationData(msexp, identFile)

## access featureData; please note the multiple identification data
## for spectrum 1 (row 1)
fData(msexp)

#>       spectrum          sequence chargeState rank passThreshold
#> F1.S1        1 VESITARHGEVLQLRPK           3    1          TRUE
#> F1.S2        2     IDGQWVTHQWLKK           3    1          TRUE
#> F1.S3        3              <NA>          NA   NA            NA
#>       experimentalMassToCharge calculatedMassToCharge modNum isDecoy post  pre
#> F1.S1                 645.3741               645.0375      0   FALSE    A    R
#> F1.S2                 546.9586               546.9633      0   FALSE    A    K
#> F1.S3                       NA                     NA     NA      NA <NA> <NA>
#>       start end  DatabaseAccess DBseqLength DatabaseSeq
#> F1.S1   170 186 ECA0984;ECA3829         231            
#> F1.S2    50  62         ECA1028         275            
#> F1.S3    NA  NA            <NA>          NA        <NA>
#>                                                                               DatabaseDescription
#> F1.S1 ECA0984 DNA mismatch repair protein;ECA3829 acetolactate synthase isozyme III large subunit
#> F1.S2                  ECA1028 2,3,4,5-tetrahydropyridine-2,6-dicarboxylate N-succinyltransferase
#> F1.S3                                                                                        <NA>
#>                idFile MS.GF.RawScore MS.GF.DeNovoScore MS.GF.SpecEValue
#> F1.S1 dummyiTRAQ.mzid            -39                77     5.527468e-05
#> F1.S2 dummyiTRAQ.mzid            -30                39     9.399048e-06
#> F1.S3            <NA>             NA                NA               NA
#>       MS.GF.EValue modName modMass modLocation subOriginalResidue
#> F1.S1     79.36958    <NA>      NA          NA               <NA>
#> F1.S2     13.46615    <NA>      NA          NA               <NA>
#> F1.S3           NA    <NA>      NA          NA               <NA>
#>       subReplacementResidue subLocation nprot npep.prot npsm.prot npsm.pep
#> F1.S1                  <NA>          NA     2         1         1        1
#> F1.S2                  <NA>          NA     1         1         1        1
#> F1.S3                  <NA>          NA    NA        NA        NA       NA
#>  [ reached getOption("max.print") -- omitted 2 rows ]

idSummary(msexp)

#>       spectrumFile          idFile coverage
#> 1 dummyiTRAQ.mzXML dummyiTRAQ.mzid      0.6