Generate an average MSnSet

Given a list of MSnSet instances, typically representing replicated experiments, the function returns an average MSnSet.

averageMSnSet(x, avg = function(x) mean(x, na.rm = TRUE), disp = npcv)

Arguments

x
A list of valid MSnSet instances to be averaged.

avg
The averaging function. Default is the mean after removing missing values, as computed by function(x) mean(x, na.rm = TRUE).

disp
The disperion function. Default is an non-parametric coefficient of variation that replaces the standard deviation by the median absolute deviation as computed by mad(x)/abs(mean(x)). See npcv for details. Note that the mad of a single value is 0 (as opposed to NA for the standard deviation, see example below).

Value

A new average MSnSet.

Details

This function is aimed at facilitating the visualisation of replicated experiments and should not be used as a replacement for a statistical analysis.

The samples of the instances to be averaged must be identical but can be in a different order (they will be reordered by default). The features names of the result will correspond to the union of the feature names of the input MSnSet instances. Each average value will be computed by the avg function and the dispersion of the replicated measurements will be estimated by the disp function. These dispersions will be stored as a data.frame in the feature metadata that can be accessed with fData(.)$disp. Similarly, the number of missing values that were present when average (and dispersion) were computed are available in fData(.)$disp.

Currently, the feature metadata of the returned object corresponds the the feature metadata of the first object in the list (augmented with the missing value and dispersion values); the metadata of the features that were missing in this first input are missing (i.e. populated with NAs). This may change in the future.

See also

Examples

library("pRolocdata")

#> Error: package ‘MSnbase’ required by ‘pRolocdata’ could not be found

## 3 replicates from Tan et al. 2009
data(tan2009r1)

#> Warning: data set ‘tan2009r1’ not found

data(tan2009r2)

#> Warning: data set ‘tan2009r2’ not found

data(tan2009r3)

#> Warning: data set ‘tan2009r3’ not found

x <- MSnSetList(list(tan2009r1, tan2009r2, tan2009r3))

#> Error in MSnSetList(list(tan2009r1, tan2009r2, tan2009r3)): object 'tan2009r1' not found

avg <- averageMSnSet(x)

#> Error in inherits(x, "MSnSetList"): object 'x' not found

dim(avg)

#> Error in eval(expr, envir, enclos): object 'avg' not found

head(exprs(avg))

#> Error in exprs(avg): object 'avg' not found

head(fData(avg)$nNA)

#> Error in fData(avg): object 'avg' not found

head(fData(avg)$disp)

#> Error in fData(avg): object 'avg' not found

## using the standard deviation as measure of dispersion
avg2 <-averageMSnSet(x, disp = sd)

#> Error in averageMSnSet(x, disp = sd): object 'x' not found

head(fData(avg2)$disp)

#> Error in fData(avg2): object 'avg2' not found

## keep only complete observations, i.e proteins 
## that had 0 missing values for all samples
sel <- apply(fData(avg)$nNA, 1 , function(x) all(x == 0))

#> Error in fData(avg): object 'avg' not found

avg <- avg[sel, ]

#> Error in eval(expr, envir, enclos): object 'avg' not found

disp <- rowMax(fData(avg)$disp)

#> Error in fData(avg): object 'avg' not found

library("pRoloc")

#> Error: package ‘MSnbase’ required by ‘pRoloc’ could not be found

setStockcol(paste0(getStockcol(), "AA"))

#> Error in setStockcol(paste0(getStockcol(), "AA")): could not find function "setStockcol"

plot2D(avg, cex = 7.7 * disp)

#> Error in plot2D(avg, cex = 7.7 * disp): could not find function "plot2D"

title(main = paste("Dispersion: non-parametric CV",
                   paste(round(range(disp), 3), collapse = " - ")))

#> Error in eval(quote(list(...)), env): object 'disp' not found