Calculate ions produced by fragmentation.

These method calculates a-, b-, c-, x-, y- and z-ions produced by fragmentation.

Arguments

sequence

character, peptide sequence.

object

Object of class "Spectrum2" or "missing" .

tolerance

numeric tolerance between the theoretical and measured MZ values (only available if object is not missing).

method

method used for for duplicated matches. Choose "highest" or "closest" to select the peak with the highest intensity respectively the closest MZ in the tolerance range. If "all" is given all possible matches in the tolerance range are reported (only available if object is not missing).

type

character vector of target ions; possible values: c("a", "b", "c", "x", "y", "z"); default: type=c("b", "y").

z

numeric desired charge state; default z=1.

modifications

named numeric vector of used modifications. The name must correspond to the one-letter-code of the modified amino acid and the numeric value must represent the mass that should be added to the original amino accid mass, default: Carbamidomethyl modifications=c(C=57.02146). Use Nterm or Cterm as names for modifications that should be added to the amino respectively carboxyl-terminus.

neutralLoss

list, it has to have two named elments, namely water and ammonia that contain a character vector which type of neutral loss should be calculated. Currently neutral loss on the C terminal "Cterm", at the amino acids c("D", "E", "S", "T") for "water" (shown with an _) and c("K", "N", "Q", "R") for "ammonia" (shown with an *) are supported. There is a helper function defaultNeutralLoss that returns the correct list. It has two arguments disableWaterLoss and disableAmmoniaLoss to remove single neutral loss options. See the example section for use cases.

verbose

logical if TRUE (default) the used modifications are printed.

Methods

signature(sequence = "character", object = "missing", …)

Calculates the theoretical fragments for a peptide sequence. Returns a data.frame with the columns c("mz", "ion", "type", "pos", "z", "seq").

signature(sequence = "character", object = "Spectrum2", …)

Calculates and matches the theoretical fragments for a peptide sequence and a "Spectrum2" object. The … arguments are passed to the internal functions. Currently tolerance, method and relative are supported. You could change the tolerance (default 25e-6) and decide whether this tolerance should be applied relative (default relative = TRUE) or absolute (relative = FALSE) to match the theoretical fragment MZ with the MZ of the spectrum. In cases of multiple matches use method to select the peak with the highest intensity (method = "highest", default) respectively closest MZ (method = "closes"). If method = "all" is set all possible matches in the current tolerance range are reported. Returns the same data.frame as above but the mz column represents the matched MZ values of the spectrum. Additionally there is a column error that contains the difference between the observed MZ (from the spectrum) to the theoretical fragment MZ.

Examples

## find path to a mzXML file
file <- dir(system.file(package = "MSnbase", dir = "extdata"),
            full.name = TRUE, pattern = "mzXML$")

## create basic MSnExp
msexp <- readMSData(file, centroided = FALSE)

#> Reading 5 MS2 spectra from file dummyiTRAQ.mzXML
#>   |                                                                              |                                                                      |   0%  |                                                                              |==============                                                        |  20%  |                                                                              |============================                                          |  40%  |                                                                              |==========================================                            |  60%  |                                                                              |========================================================              |  80%  |                                                                              |======================================================================| 100%
#> Creating 'MSnExp' object

## centroid them
msexp <- pickPeaks(msexp)

#>   |                                                                              |                                                                      |   0%  |                                                                              |==============                                                        |  20%  |                                                                              |============================                                          |  40%  |                                                                              |==========================================                            |  60%  |                                                                              |========================================================              |  80%  |                                                                              |======================================================================| 100%

## calculate fragments for ACE with default modification
calculateFragments("ACE", modifications=c(C=57.02146))

#> Modifications used: C=57.02146

#>          mz ion type pos z seq
#> 1  72.04439  b1    b   1 1   A
#> 2 232.07504  b2    b   2 1  AC
#> 3 361.11763  b3    b   3 1 ACE
#> 4 148.06043  y1    y   1 1   E
#> 5 308.09108  y2    y   2 1  CE
#> 6 379.12819  y3    y   3 1 ACE
#> 7 115.06278 y1_   y_   1 1   E
#> 8 275.09343 y2_   y_   2 1  CE
#> 9 346.13054 y3_   y_   3 1 ACE

## calculate fragments for ACE with an addition N-terminal modification
calculateFragments("ACE", modifications=c(C=57.02146, Nterm=229.1629))

#> Modifications used: C=57.02146, Nterm=229.1629

#>         mz ion type pos z seq
#> 1 301.2073  b1    b   1 1   A
#> 2 461.2379  b2    b   2 1  AC
#> 3 590.2805  b3    b   3 1 ACE
#> 4 148.0604  y1    y   1 1   E
#> 5 308.0911  y2    y   2 1  CE
#> 6 379.1282  y3    y   3 1 ACE
#> 7 115.0628 y1_   y_   1 1   E
#> 8 275.0934 y2_   y_   2 1  CE
#> 9 346.1305 y3_   y_   3 1 ACE

## calculate fragments for ACE without any modifications
calculateFragments("ACE", modifications=NULL)

#> Modifications used: None

#>          mz ion type pos z seq
#> 1  72.04439  b1    b   1 1   A
#> 2 175.05358  b2    b   2 1  AC
#> 3 304.09617  b3    b   3 1 ACE
#> 4 148.06043  y1    y   1 1   E
#> 5 251.06962  y2    y   2 1  CE
#> 6 322.10673  y3    y   3 1 ACE
#> 7 115.06278 y1_   y_   1 1   E
#> 8 218.07197 y2_   y_   2 1  CE
#> 9 289.10908 y3_   y_   3 1 ACE

calculateFragments("VESITARHGEVLQLRPK",
                   type=c("a", "b", "c", "x", "y", "z"),
                   z=1:2)

#> Modifications used: C=57.02146

#>             mz  ion type pos z               seq
#> 1     72.08077   a1    a   1 1                 V
#> 2     36.54402   a1    a   1 2                 V
#> 3    201.12336   a2    a   2 1                VE
#> 4    101.06532   a2    a   2 2                VE
#> 5    288.15539   a3    a   3 1               VES
#> 6    144.58133   a3    a   3 2               VES
#> 7    401.23945   a4    a   4 1              VESI
#> 8    201.12336   a4    a   4 2              VESI
#> 9    502.28713   a5    a   5 1             VESIT
#> 10   251.64720   a5    a   5 2             VESIT
#> 11   573.32424   a6    a   6 1            VESITA
#> 12   287.16576   a6    a   6 2            VESITA
#> 13   729.42535   a7    a   7 1           VESITAR
#> 14   365.21631   a7    a   7 2           VESITAR
#> 15   866.48426   a8    a   8 1          VESITARH
#> 16   433.74577   a8    a   8 2          VESITARH
#>  [ reached getOption("max.print") -- omitted 518 rows ]

calculateFragments("VESITARHGEVLQLRPK", msexp[[1]])

#> Modifications used: C=57.02146

#>          mz intensity  ion type pos z               seq       error
#> 1  114.1109  706555.7  y1_   y_   1 1                 K  0.00425275
#> 2  429.2563 1972344.0   b4    b   4 1              VESI -0.02189010
#> 3  513.3047 2574137.0   y4    y   4 1              LRPK  0.04598246
#> 4  754.4504  537234.8   y6    y   6 1            LQLRPK  0.04293155
#> 5  982.5354  500159.1   y8    y   8 1          EVLQLRPK  0.06897061
#> 6 1080.5867  209363.7  b10    b  10 1        VESITARHGE -0.04344392
#> 7 1688.0375  136748.8 y15*   y*  15 1   SITARHGEVLQLRPK -0.07729359
#> 8 1882.0074  149649.1 b17_   b_  17 1 VESITARHGEVLQLRPK  0.08206471

## neutral loss
defaultNeutralLoss()

#> $water
#> [1] "Cterm" "D"     "E"     "S"     "T"    
#> 
#> $ammonia
#> [1] "K" "N" "Q" "R"
#> 

## disable water loss on the C terminal
defaultNeutralLoss(disableWaterLoss="Cterm")

#> $water
#> [1] "D" "E" "S" "T"
#> 
#> $ammonia
#> [1] "K" "N" "Q" "R"
#> 

## real example
calculateFragments("PQR")

#> Modifications used: C=57.02146

#>           mz ion type pos z seq
#> 1   98.06004  b1    b   1 1   P
#> 2  226.11862  b2    b   2 1  PQ
#> 3  382.21973  b3    b   3 1 PQR
#> 4  175.11895  y1    y   1 1   R
#> 5  303.17753  y2    y   2 1  QR
#> 6  400.23029  y3    y   3 1 PQR
#> 7  142.12130 y1_   y_   1 1   R
#> 8  270.17988 y2_   y_   2 1  QR
#> 9  367.23264 y3_   y_   3 1 PQR
#> 10 365.19318 b3*   b*   3 1 PQR
#> 11 286.15098 y2*   y*   2 1  QR
#> 12 383.20374 y3*   y*   3 1 PQR

calculateFragments("PQR",
                   neutralLoss=defaultNeutralLoss(disableWaterLoss="Cterm"))

#> Modifications used: C=57.02146

#>          mz ion type pos z seq
#> 1  98.06004  b1    b   1 1   P
#> 2 226.11862  b2    b   2 1  PQ
#> 3 382.21973  b3    b   3 1 PQR
#> 4 175.11895  y1    y   1 1   R
#> 5 303.17753  y2    y   2 1  QR
#> 6 400.23029  y3    y   3 1 PQR
#> 7 365.19318 b3*   b*   3 1 PQR
#> 8 286.15098 y2*   y*   2 1  QR
#> 9 383.20374 y3*   y*   3 1 PQR

calculateFragments("PQR",
                   neutralLoss=defaultNeutralLoss(disableAmmoniaLoss="Q"))

#> Modifications used: C=57.02146

#>          mz ion type pos z seq
#> 1  98.06004  b1    b   1 1   P
#> 2 226.11862  b2    b   2 1  PQ
#> 3 382.21973  b3    b   3 1 PQR
#> 4 175.11895  y1    y   1 1   R
#> 5 303.17753  y2    y   2 1  QR
#> 6 400.23029  y3    y   3 1 PQR
#> 7 142.12130 y1_   y_   1 1   R
#> 8 270.17988 y2_   y_   2 1  QR
#> 9 367.23264 y3_   y_   3 1 PQR

## disable neutral loss completely
calculateFragments("PQR", neutralLoss=NULL)

#> Modifications used: C=57.02146

#>          mz ion type pos z seq
#> 1  98.06004  b1    b   1 1   P
#> 2 226.11862  b2    b   2 1  PQ
#> 3 382.21973  b3    b   3 1 PQR
#> 4 175.11895  y1    y   1 1   R
#> 5 303.17753  y2    y   2 1  QR
#> 6 400.23029  y3    y   3 1 PQR