getVariableName {MSnbase}	R Documentation

Return a variable name

Description

Return the name of variable varname in call match_call.

Usage

getVariableName(match_call, varname)

Arguments

match_call
An object of class call, as returned by match.call.

varname
An character of length 1 which is looked up in match_call.

Value

A character with the name of the variable passed as parameter varname in parent close of match_call.

Author(s)

Laurent Gatto

Examples

a <- 1
f <- function(x, y)
 MSnbase:::getVariableName(match.call(), "x")
f(x = a)
f(y = a)

[Package MSnbase version 2.10.1 Index]