hasSpectra {MSnbase}	R Documentation

Checks if raw data files have any spectra or chromatograms

Description

Helper functions to check whether raw files contain spectra or chromatograms.

Usage

hasSpectra(files)

hasChromatograms(files)

Arguments

files
A character() with raw data filenames.

Value

A logical(n) where n == length(x) with TRUE if that files contains at least one spectrum, FALSE otherwise.

Author(s)

Laurent Gatto

Examples

f <- msdata::proteomics(full.names = TRUE)[1:2]
hasSpectra(f)
hasChromatograms(f)

[Package MSnbase version 2.12.0 Index]