makeCamelCase {MSnbase}	R Documentation

Convert to camel case by replacing dots by captial letters

Description

Convert a vector of characters to camel case by replacing dots by captial letters.

Usage

makeCamelCase(x, prefix)

Arguments

x
A vector to be transformed to camel case.

prefix
An optional character of length one. Any additional elements are ignores.

Value

A character of same length as x.

Author(s)

Laurent Gatto

Examples

nms <- c("aa.foo", "ab.bar")
makeCamelCase(nms)
makeCamelCase(nms, prefix = "x")

[Package MSnbase version 2.4.2 Index]