How many features in a group?

This function computes the number of features in the group defined by the feature variable fcol and appends this information in the feature data of object.

nFeatures(object, fcol)

Arguments

object

An instance of class MSnSet.

fcol

Feature variable defining the feature grouping structure.

Value

An updated MSnSet with a new feature variable fcol.nFeatures.

Examples

library(pRolocdata)

#> Error: package ‘MSnbase’ required by ‘pRolocdata’ could not be found

data("hyperLOPIT2015ms3r1psm")

#> Warning: data set ‘hyperLOPIT2015ms3r1psm’ not found

hyperLOPIT2015ms3r1psm <- nFeatures(hyperLOPIT2015ms3r1psm,
                                    "Protein.Group.Accessions")

#> Error in inherits(object, "MSnSet"): object 'hyperLOPIT2015ms3r1psm' not found

i <- c("Protein.Group.Accessions", "Protein.Group.Accessions.nFeatures")
fData(hyperLOPIT2015ms3r1psm)[1:10, i]

#> Error in fData(hyperLOPIT2015ms3r1psm): object 'hyperLOPIT2015ms3r1psm' not found