plot2d-methods {MSnbase}	R Documentation

The 'plot2d' method for 'MSnExp' quality assessment

Description

These methods plot the retention time vs. precursor MZ for the whole "MSnExp" experiment. Individual dots will be colour-coded to describe individual spectra's peaks count, total ion count, precursor charge (MS2 only) or file of origin.

The methods make use the ggplot2 system. An object of class 'ggplot' is returned invisibly.

Arguments

object
An object of class "MSnExp" or a data.frame. In the latter case, the data frame must have numerical columns named 'retention.time' and 'precursor.mz' and one of 'tic', 'file', 'peaks.count' or 'charge', depending on the z parameter. Such a data frame is typically generated using the header method on "MSnExp" object.

z
A character indicating according to what variable to colour the dots. One of, possibly abreviated, "ionCount" (total ion count), "file" (raw data file), "peaks.count" (peaks count) or "charge" (precursor charge).

alpha
Numeric [0,1] indicating transparence level of points.

plot
A logical indicating whether the plot should be printed (default is 'TRUE').

Methods

signature(object = "MSnExp", ...)

Plots a 'MSnExp' summary.

signature(object = "data.frame", ...)

Plots a summary of the 'MSnExp' experiment described by the data frame.

Author(s)

Laurent Gatto <lg390@cam.ac.uk>

See Also

The plotDensity and plotMzDelta methods for other QC plots.

Examples

itraqdata
plot2d(itraqdata,z="ionCount")
plot2d(itraqdata,z="peaks.count")
plot2d(itraqdata,z="charge")

[Package MSnbase version 2.12.0 Index]