plotDensity-methods {MSnbase}	R Documentation

The 'plotDensity' method for 'MSnExp' quality assessment

Description

These methods plot the distribution of several parameters of interest for the different precursor charges for "MSnExp" experiment.

The methods make use the ggplot2 system. An object of class 'ggplot' is returned invisibly.

Arguments

object
An object of class "MSnExp" or and 'data.frame'. In the latter case, the data frame must have numerical columns named 'charge' and one of 'precursor.mz', 'peaks.count' or 'ionCount', depending on the z parameter. Such a data frame is typically generated using the header method on "MSnExp" object.

z
A character indicating which parameter's densitiy to plot. One of, possibly abreviated, "ionCount" (total ion count), "peaks.count" (peaks count) or "precursor.mz" (precursor MZ).

log
Logical, whether to log transform the data (default is 'FALSE').

plot
A logical indicating whether the plot should be printed (default is 'TRUE').

Methods

signature(object = "MSnExp", ...)

Plots a 'MSnExp' summary.

signature(object = "data.frame", ...)

Plots a summary of the 'MSnExp' experiment described by the data frame.

Author(s)

Laurent Gatto <lg390@cam.ac.uk>

See Also

The plot2d and plotDensity methods for other QC plots.

Examples

itraqdata
plotDensity(itraqdata,z="ionCount")
plotDensity(itraqdata,z="peaks.count")
plotDensity(itraqdata,z="precursor.mz")

[Package MSnbase version 2.4.2 Index]