Adds Identification Data

This methods add identification data to an experiment "MSnExp" or to a "MSnSet".

Methods

signature(object = "MSnExp", id = "character", fcol = "character", icol = "character", verbose = "logical")

Adds the identification data stored in mzIdentML files to a "MSnExp" instance. The method handles one or multiple mzIdentML files provided via id. id has to be a character vector of valid filenames. fcol and icol specify the columns in the featureData slot and the identification data.frame that are used to merge both data together as character vectors (defaults are fcol = c("spectrum.file", "acquisition.number") and icol = c("spectrumFile", "acquisitionnum")). The verbose argument (default is TRUE) defines whether status messages should be showed.

signature(object = "MSnExp", id = "mzID", fcol = "character", icol = "character", verbose = "logical")

Same as above but id could be an mzID object.

signature(object = "MSnExp", id = "mzIDCollection", fcol = "character", icol = "character", verbose = "logical")

Same as above but id could be an mzIDCollection object.

signature(object = "MSnExp", id = "data.frame", fcol = "character", icol = "character", verbose = "logical")

Same as above but id could be a data.frame.

signature(object = "MSnSet", id = "character", verbose = "logical")

Adds the identification data stored in mzIdentML files to an "MSnSet" instance. The method handles one or multiple mzIdentML files provided via id. id has to be a character vector of valid filenames.

fcol and icol specify the columns in the featureData slot and the identification data.frame that are used to merge both data together as character vectors (defaults are fcol = c("spectrum.file", "acquisition.number") and icol = c("spectrumFile", "acquisitionnum")).

The verbose argument (default is TRUE) defines whether status messages should be showed.

signature(object = "MSnSet", id = "mzID", fcol = "character", icol = "character", verbose = "logical")

Same as above but id could be an mzID object.

signature(object = "MSnSet", id = "mzIDCollection", fcol = "character", icol = "character", verbose = "logical")

Same as above but id could be an mzIDCollection object.

signature(object = "MSnSet", id = "data.frame", fcol = "character", icol = "character", verbose = "logical")

Same as above but id could be a data.frame.

Details

The featureData slots in an "MSnExp" or an "MSnSet" instance provides only one row per MS2 spectrum but the identification is not always bijective. If multiple possible matches are present only the highest ranked identification is added.

The column nprot contains the number of members in the protein group; the columns accession and description contain a semicolon separated list of all matches sorted by their rank values. The columns npsm.prot and npep.prot represent the number of PSMs and peptides that were matched to a particular protein group. The column npsm.pep indicates how many PSMs were attributed to a peptide (as defined by its sequence pepseq).

See also

MSnExp and MSnSet.

Examples

## find path to a mzXML file
quantFile <- dir(system.file(package = "MSnbase", dir = "extdata"),
                 full.name = TRUE, pattern = "mzXML$")
## find path to a mzIdentML file
identFile <- dir(system.file(package = "MSnbase", dir = "extdata"),
                 full.name = TRUE, pattern = "dummyiTRAQ.mzid")

## create basic MSnExp
msexp <- readMSData(quantFile)

#> Reading 5 MS2 spectra from file dummyiTRAQ.mzXML
#>   |                                                                              |                                                                      |   0%  |                                                                              |==============                                                        |  20%  |                                                                              |============================                                          |  40%  |                                                                              |==========================================                            |  60%  |                                                                              |========================================================              |  80%  |                                                                              |======================================================================| 100%
#> Creating 'MSnExp' object

## add identification information
msexp <- addIdentificationData(msexp, identFile)

#> reading dummyiTRAQ.mzid... DONE!

## access featureData; please note the multiple identification data
## for spectrum 1 (row 1)
fData(msexp)

#>      spectrum scan number(s) passthreshold rank calculatedmasstocharge
#> X1.1        1              1          TRUE    1               645.0375
#> X2.1        2              2          TRUE    1               546.9633
#> X3.1        3             NA            NA   NA                     NA
#>      experimentalmasstocharge chargestate ms-gf:denovoscore ms-gf:evalue
#> X1.1                 645.3741           3                77     79.36958
#> X2.1                 546.9586           3                39     13.46615
#> X3.1                       NA          NA                NA           NA
#>      ms-gf:rawscore ms-gf:specevalue assumeddissociationmethod isotopeerror
#> X1.1            -39     5.527468e-05                       CID            1
#> X2.1            -30     9.399048e-06                       CID            0
#> X3.1             NA               NA                      <NA>         <NA>
#>      isdecoy post  pre end start       accession length
#> X1.1   FALSE    A    R 186   170 ECA0984;ECA3829    231
#> X2.1   FALSE    A    K  62    50         ECA1028    275
#> X3.1      NA <NA> <NA>  NA    NA            <NA>     NA
#>                                                                      description
#> X1.1 DNA mismatch repair protein;acetolactate synthase isozyme III large subunit
#> X2.1          2,3,4,5-tetrahydropyridine-2,6-dicarboxylate N-succinyltransferase
#> X3.1                                                                        <NA>
#>                 pepseq modified modification          idFile
#> X1.1 VESITARHGEVLQLRPK    FALSE           NA dummyiTRAQ.mzid
#> X2.1     IDGQWVTHQWLKK    FALSE           NA dummyiTRAQ.mzid
#> X3.1              <NA>       NA           NA            <NA>
#>                  databaseFile nprot npep.prot npsm.prot npsm.pep
#> X1.1 erwinia_carotovora.fasta     2         1         1        1
#> X2.1 erwinia_carotovora.fasta     1         1         1        1
#> X3.1                     <NA>    NA        NA        NA       NA
#>  [ reached getOption("max.print") -- omitted 2 rows ]

idSummary(msexp)

#>       spectrumFile          idFile coverage
#> 1 dummyiTRAQ.mzXML dummyiTRAQ.mzid      0.6