Keep only common feature names

Subsets MSnSet instances to their common feature names.

commonFeatureNames(x, y)

Arguments

x
An instance of class MSnSet or a list or MSnSetList with at least 2 MSnSet objects.

y
An instance of class MSnSet. Ignored if x is a list/MSnSetList.

Value

An linkS4class{MSnSetList} composed of the input MSnSet containing only common features in the same order. The names of the output are either the names of the x and y input variables or the names of x if a list is provided.

Examples

library("pRolocdata")

#> Error: package ‘MSnbase’ required by ‘pRolocdata’ could not be found

data(tan2009r1)

#> Warning: data set ‘tan2009r1’ not found

data(tan2009r2)

#> Warning: data set ‘tan2009r2’ not found

cmn <- commonFeatureNames(tan2009r1, tan2009r2)

#> Error in inherits(x, "MSnSetList"): object 'tan2009r1' not found

names(cmn)

#> Error in eval(expr, envir, enclos): object 'cmn' not found

## as a named list
names(commonFeatureNames(list(a = tan2009r1, b = tan2009r2)))

#> Error in commonFeatureNames(list(a = tan2009r1, b = tan2009r2)): object 'tan2009r1' not found

## without message
suppressMessages(cmn <- commonFeatureNames(tan2009r1, tan2009r2))

#> Error in commonFeatureNames(tan2009r1, tan2009r2): object 'tan2009r1' not found

## more than 2 instance
data(tan2009r3)

#> Warning: data set ‘tan2009r3’ not found

cmn <- commonFeatureNames(list(tan2009r1, tan2009r2, tan2009r3))

#> Error in commonFeatureNames(list(tan2009r1, tan2009r2, tan2009r3)): object 'tan2009r1' not found

length(cmn)

#> Error in eval(expr, envir, enclos): object 'cmn' not found