Expand or merge feature variables

The expandFeatureVars and mergeFeatureVars respectively expand and merge groups of feature variables. Using these functions, a set of columns in a feature data can be merged into a single new data.frame-column variables and a data.frame-column can be expanded into single feature columns. The original feature variables are removed.

expandFeatureVars(x, fcol, prefix)

mergeFeatureVars(x, fcol, fcol2)

Arguments

x
An object of class MSnSet.

fcol
A character() of feature variables to expand (for expandFeatureVars) or merge (for mergeFeatureVars).

prefix
A character(1) to use as prefix to the new feature variables. If missing (default), then fcol is used instead. If NULL, then no prefix is used.

fcol2
A character(1) defining the name of the new feature variable.

Value

An MSnSet for expanded (merged) feature variables.

Examples

library("pRolocdata")

#> Error: package ‘MSnbase’ required by ‘pRolocdata’ could not be found

data(hyperLOPIT2015)

#> Warning: data set ‘hyperLOPIT2015’ not found

fvarLabels(hyperLOPIT2015)

#> Error in fvarLabels(hyperLOPIT2015): object 'hyperLOPIT2015' not found

## Let's merge all svm prediction feature variables
(k <- grep("^svm", fvarLabels(hyperLOPIT2015), value = TRUE))

#> Error in fvarLabels(hyperLOPIT2015): object 'hyperLOPIT2015' not found

hl <- mergeFeatureVars(hyperLOPIT2015, fcol = k, fcol2 = "SVM")

#> Error in fvarLabels(x): object 'hyperLOPIT2015' not found

fvarLabels(hl)

#> Error in fvarLabels(hl): object 'hl' not found

head(fData(hl)$SVM)

#> Error in fData(hl): object 'hl' not found

## Let's expand the new SVM into individual columns
hl2 <- expandFeatureVars(hl, "SVM")

#> Error in is.data.frame(fData(x)): object 'hl' not found

fvarLabels(hl2)

#> Error in fvarLabels(hl2): object 'hl2' not found

## We can set the prefix manually
hl2 <- expandFeatureVars(hl, "SVM", prefix = "Expanded")

#> Error in is.data.frame(fData(x)): object 'hl' not found

fvarLabels(hl2)

#> Error in fvarLabels(hl2): object 'hl2' not found

## If we don't want any prefix
hl2 <- expandFeatureVars(hl, "SVM", prefix = NULL)

#> Error in is.data.frame(fData(x)): object 'hl' not found

fvarLabels(hl2)

#> Error in fvarLabels(hl2): object 'hl2' not found