Calculates coeffivient of variation for features

This function calculates the column-wise coefficient of variation (CV), i.e. the ration between the standard deviation and the mean, for the features in an "MSnSet". The CVs are calculated for the groups of features defined by groupBy. For groups defined by single features, NA is returned.

featureCV(x, groupBy, na.rm = TRUE, norm = c("sum", "max", "none",
  "center.mean", "center.median", "quantiles", "quantiles.robust"))

Arguments

x

An instance of class "MSnSet".

groupBy

An object of class factor defining how to summerise the features.

na.rm

A logical defining whether missing values should be removed.

norm

One of 'none' (default), 'sum', 'max', 'center.mean', 'center.median' 'quantiles' or 'quantiles.robust' defining if and how the data should be normalised prior to CV calculation. See normalise for more details.

Value

A matrix of dimensions length(levels(groupBy)) by ncol(x) with the respecive CVs.

See also

combineFeatures

Examples

data(msnset)
msnset <- msnset[1:4]
gb <- factor(rep(1:2, each = 2))
featureCV(msnset, gb)

#>   CV.iTRAQ4.114 CV.iTRAQ4.115 CV.iTRAQ4.116 CV.iTRAQ4.117
#> 1     0.6310418     0.3633932  0.1041115783     0.3714580
#> 2     0.0113106     0.1072691  0.0005836238     0.1240521