iPQF {MSnbase}	R Documentation

iPQF: iTRAQ (and TMT) Protein Quantification based on Features

Description

The iPQF spectra-to-protein summarisation method integrates peptide spectra characteristics and quantitative values for protein quantitation estimation. Spectra features, such as charge state, sequence length, identification score and others, contain valuable information concerning quantification accuracy. The iPQF algorithm assigns weights to spectra according to their overall feature reliability and computes a weighted mean to estimate protein quantities. See also combineFeatures for a more general overview of feature aggregation and examples.

Usage

iPQF(object, groupBy, low.support.filter = FALSE, ratio.calc = "sum",
  method.combine = FALSE, feature.weight = c(7, 6, 4, 3, 2, 1, 5)^2)

Arguments

object
An instance of class MSnSet containing absolute ion intensities.

groupBy
Vector defining spectra to protein matching. Generally, this is a feature variable such as fData(object)$accession.

low.support.filter
A logical specifying if proteins being supported by only 1-2 peptide spectra should be filtered out. Default is FALSE.

ratio.calc
Either "none" (don't calculate any ratios), "sum" (default), or a specific channel (one of sampleNames(object)) defining how to calculate relative peptides intensities.

method.combine
A logical defining whether to further use median polish to combine features.

feature.weight
Vector "numeric" giving weight to the different features. Default is the squared order of the features redundant -unique-distance metric, charge state, ion intensity, sequence length, identification score, modification state, and mass based on a robustness analysis.

Value

A matrix with estimated protein ratios.

Author(s)

Martina Fischer

References

iPQF: a new peptide-to-protein summarization method using peptide spectra characteristics to improve protein quantification. Fischer M, Renard BY. Bioinformatics. 2016 Apr 1;32(7):1040-7. doi:10.1093/bioinformatics/btv675. Epub 2015 Nov 20. PubMed PMID:26589272.

Examples

data(msnset2)
head(exprs(msnset2))
prot <- combineFeatures(msnset2,
                        groupBy = fData(msnset2)$accession,
                        fun = "iPQF")
head(exprs(prot))

[Package MSnbase version 2.4.2 Index]