Create a data with missing values

These functions take an instance of class "MSnSet" and sets randomly selected values to NA.

makeNaData(object, nNA, pNA, exclude)

makeNaData2(object, nRows, nNAs, exclude)

whichNA(x)

Arguments

object
An instance of class MSnSet.

nNA
The absolute number of missing values to be assigned.

pNA
The proportion of missing values to be assignmed.

exclude
A vector to be used to subset object, defining rows that should not be used to set NAs.

nRows
The number of rows for each set.

nNAs
The number of missing values for each set.

x
A matrix or an instance of class MSnSet.

Value

An instance of class MSnSet, as object, but with the appropriate number/proportion of missing values. The returned object has an additional feature meta-data columns, nNA

Details

makeNaData randomly selects a number nNA (or a proportion pNA) of cells in the expression matrix to be set to NA.

makeNaData2 will select length(nRows) sets of rows from object, each with nRows[i] rows respectively. The first set will be assigned nNAs[1] missing values, the second nNAs[2], ... As opposed to makeNaData, this permits to control the number of NAs per rows.

The whichNA can be used to extract the indices of the missing values, as illustrated in the example.

Examples

## Example 1
library(pRolocdata)

#> Error: package ‘MSnbase’ required by ‘pRolocdata’ could not be found

data(dunkley2006)

#> Warning: data set ‘dunkley2006’ not found

sum(is.na(dunkley2006))

#> Error in eval(expr, envir, enclos): object 'dunkley2006' not found

dunkleyNA <- makeNaData(dunkley2006, nNA = 150)

#> Error in inherits(object, "MSnSet"): object 'dunkley2006' not found

processingData(dunkleyNA)

#> Error in processingData(dunkleyNA): object 'dunkleyNA' not found

sum(is.na(dunkleyNA))

#> Error in eval(expr, envir, enclos): object 'dunkleyNA' not found

table(fData(dunkleyNA)$nNA)

#> Error in fData(dunkleyNA): object 'dunkleyNA' not found

naIdx <- whichNA(dunkleyNA)

#> Error in inherits(x, "MSnSet"): object 'dunkleyNA' not found

head(naIdx)

#> Error in head(naIdx): object 'naIdx' not found

## Example 2
dunkleyNA <- makeNaData(dunkley2006, nNA = 150, exclude = 1:10)

#> Error in inherits(object, "MSnSet"): object 'dunkley2006' not found

processingData(dunkleyNA)

#> Error in processingData(dunkleyNA): object 'dunkleyNA' not found

table(fData(dunkleyNA)$nNA[1:10])

#> Error in fData(dunkleyNA): object 'dunkleyNA' not found

table(fData(dunkleyNA)$nNA)

#> Error in fData(dunkleyNA): object 'dunkleyNA' not found

## Example 3
nr <- rep(10, 5)
na <- 1:5
x <- makeNaData2(dunkley2006[1:100, 1:5],
                 nRows = nr,
                 nNAs = na)

#> Error in inherits(object, "MSnSet"): object 'dunkley2006' not found

processingData(x)

#> Error in processingData(x): object 'x' not found

(res <- table(fData(x)$nNA))

#> Error in fData(x): object 'x' not found

stopifnot(as.numeric(names(res)[-1]) ==  na)

#> Error in as.numeric(names(res)[-1]) == na: object 'res' not found

stopifnot(res[-1] ==  nr)

#> Error in res[-1] == nr: object 'res' not found

## Example 3
nr2 <- c(5, 12, 11, 8)
na2 <- c(3, 8, 1, 4)
x2 <- makeNaData2(dunkley2006[1:100, 1:10],
                  nRows = nr2,
                  nNAs = na2)

#> Error in inherits(object, "MSnSet"): object 'dunkley2006' not found

processingData(x2)

#> Error in processingData(x2): object 'x2' not found

(res2 <- table(fData(x2)$nNA))

#> Error in fData(x2): object 'x2' not found

stopifnot(as.numeric(names(res2)[-1]) ==  sort(na2))

#> Error in as.numeric(names(res2)[-1]) == sort(na2): object 'res2' not found

stopifnot(res2[-1] ==  nr2[order(na2)])

#> Error in res2[-1] == nr2[order(na2)]: object 'res2' not found

## Example 5
nr3 <- c(5, 12, 11, 8)
na3 <- c(3, 8, 1, 3)
x3 <- makeNaData2(dunkley2006[1:100, 1:10],
                  nRows = nr3,
                  nNAs = na3)

#> Error in inherits(object, "MSnSet"): object 'dunkley2006' not found

processingData(x3)

#> Error in processingData(x3): object 'x3' not found

(res3 <- table(fData(x3)$nNA))

#> Error in fData(x3): object 'x3' not found