Documenting missing data visualisation

There is a need for adequate handling of missing value impuation in quantitative proteomics. Before developing a framework to handle missing data imputation optimally, we propose a set of visualisation tools. This document serves as an internal notebook for current progress and ideas that will eventually materialise in exported functionality in the MSnbase package.

Details

The explore the structure of missing values, we propose to

1. Explore missing values in the frame of the experimental design. The imageNA2 function offers such a simple visualisation. It is currently limited to 2-group designs/comparisons. In case of time course experiments or sub-cellular fractionation along a density gradient, we propose to split the time/gradient into 2 groups (early/late, top/bottom) as a first approximation.

2. Explore the proportion of missing values in each group.

3. Explore the total and group-wise feature intensity distributions.

The existing plotNA function illustrates the completeness/missingness of the data.

See also

Examples

## Other suggestions
library("pRolocdata")

#> Error: package ‘MSnbase’ required by ‘pRolocdata’ could not be found

library("pRoloc")

#> Error: package ‘MSnbase’ required by ‘pRoloc’ could not be found

data(dunkley2006)

#> Warning: data set ‘dunkley2006’ not found

set.seed(1)
nax <- makeNaData(dunkley2006, pNA = 0.10)

#> Error in inherits(object, "MSnSet"): object 'dunkley2006' not found

pcol <- factor(ifelse(dunkley2006$fraction <= 5, "A", "B"))

#> Error in ifelse(dunkley2006$fraction <= 5, "A", "B"): object 'dunkley2006' not found

sel1 <- pcol == "A"

#> Error in eval(expr, envir, enclos): object 'pcol' not found

## missing values in each sample
barplot(colSums(is.na(nax)), col = pcol)

#> Error in colSums(is.na(nax)): object 'nax' not found

## table of missing values in proteins
par(mfrow = c(3, 1))
barplot(table(rowSums(is.na(nax))), main = "All")

#> Error in rowSums(is.na(nax)): object 'nax' not found

barplot(table(rowSums(is.na(nax)[sel1,])), main = "Group A")

#> Error in rowSums(is.na(nax)[sel1, ]): object 'nax' not found

barplot(table(rowSums(is.na(nax)[!sel1,])), main = "Group B")

#> Error in rowSums(is.na(nax)[!sel1, ]): object 'nax' not found

fData(nax)$nNA1 <- rowSums(is.na(nax)[, sel1])

#> Error in rowSums(is.na(nax)[, sel1]): object 'nax' not found

fData(nax)$nNA2 <- rowSums(is.na(nax)[, !sel1])

#> Error in rowSums(is.na(nax)[, !sel1]): object 'nax' not found

fData(nax)$nNA <- rowSums(is.na(nax))

#> Error in rowSums(is.na(nax)): object 'nax' not found

o <- MSnbase:::imageNA2(nax, pcol)

#> Error in MSnbase:::imageNA2(nax, pcol): object 'pcol' not found

plot((fData(nax)$nNA1 - fData(nax)$nNA2)[o], type = "l")

#> Error in fData(nax): object 'nax' not found

grid()

#> Error in int_abline(a = a, b = b, h = h, v = v, untf = untf, ...): plot.new has not been called yet

plot(sort(fData(nax)$nNA1 - fData(nax)$nNA2), type = "l")

#> Error in fData(nax): object 'nax' not found

grid()

#> Error in int_abline(a = a, b = b, h = h, v = v, untf = untf, ...): plot.new has not been called yet

o2 <- order(fData(nax)$nNA1 - fData(nax)$nNA2)

#> Error in fData(nax): object 'nax' not found

MSnbase:::imageNA2(nax, pcol, Rowv=o2)

#> Error in MSnbase:::imageNA2(nax, pcol, Rowv = o2): object 'pcol' not found

layout(matrix(c(rep(1, 10), rep(2, 5)), nc = 3))
MSnbase:::imageNA2(nax, pcol, Rowv=o2)

#> Error in MSnbase:::imageNA2(nax, pcol, Rowv = o2): object 'pcol' not found

plot((fData(nax)$nNA1 - fData(nax)$nNA)[o2], type = "l", col = "red",
     ylim = c(-9, 9), ylab = "")

#> Error in fData(nax): object 'nax' not found

lines((fData(nax)$nNA - fData(nax)$nNA2)[o2], col = "steelblue")

#> Error in fData(nax): object 'nax' not found

lines((fData(nax)$nNA1 - fData(nax)$nNA2)[o2], type = "l",
     lwd = 2)

#> Error in fData(nax): object 'nax' not found