Coerce identification data to a data.frame

A function to convert the identification data contained in an mzRident object to a data.frame. Each row represents a scan, which can however be repeated several times if the PSM matches multiple proteins and/or contains two or more modifications. To reduce the data.frame so that rows/scans are unique and use semicolon-separated values to combine information pertaining a scan, use reduce.

Arguments

from
An object of class mzRident defined in the mzR package.

Value

A data.frame

Details

See also the Tandem MS identification data section in the MSnbase-demo vignette.

Examples

## find path to a mzIdentML file
identFile <- dir(system.file(package = "MSnbase", dir = "extdata"),
                 full.name = TRUE, pattern = "dummyiTRAQ.mzid")
library("mzR")
x <- openIDfile(identFile)
x

#> Identification file handle.
#> Filename:  dummyiTRAQ.mzid 
#> Number of psms:  4 

as(x, "data.frame")

#>            sequence spectrumID chargeState rank passThreshold
#> 1     IDGQWVTHQWLKK     scan=2           3    1          TRUE
#> 2 VESITARHGEVLQLRPK     scan=1           3    1          TRUE
#> 3 IKPQAVIETLHRLTEGK     scan=1           3    2          TRUE
#>   experimentalMassToCharge calculatedMassToCharge modNum isDecoy post pre start
#> 1                 546.9586               546.9633      0   FALSE    A   K    50
#> 2                 645.3741               645.0375      0   FALSE    A   R   170
#> 3                 645.3741               645.0458      0   FALSE    A   K   372
#>   end DatabaseAccess DBseqLength DatabaseSeq
#> 1  62        ECA1028         275            
#> 2 186        ECA0984         231            
#> 3 388        ECA3829         572            
#>                                                          DatabaseDescription
#> 1 ECA1028 2,3,4,5-tetrahydropyridine-2,6-dicarboxylate N-succinyltransferase
#> 2                                        ECA0984 DNA mismatch repair protein
#> 3                    ECA3829 acetolactate synthase isozyme III large subunit
#>   acquisitionNum     spectrumFile          idFile MS.GF.RawScore
#> 1              2 dummyiTRAQ.mzXML dummyiTRAQ.mzid            -30
#> 2              1 dummyiTRAQ.mzXML dummyiTRAQ.mzid            -39
#> 3              1 dummyiTRAQ.mzXML dummyiTRAQ.mzid            -39
#>   MS.GF.DeNovoScore MS.GF.SpecEValue MS.GF.EValue modName modMass modLocation
#> 1                39     9.399048e-06     13.46615    <NA>      NA          NA
#> 2                77     5.527468e-05     79.36958    <NA>      NA          NA
#> 3                77     5.527468e-05     79.36958    <NA>      NA          NA
#>   subOriginalResidue subReplacementResidue subLocation
#> 1               <NA>                  <NA>          NA
#> 2               <NA>                  <NA>          NA
#> 3               <NA>                  <NA>          NA
#>  [ reached getOption("max.print") -- omitted 1 row ]