as {MSnbase}	R Documentation

Coerce identification data to a data.frame

Description

A function to convert the identification data contained in an mzRident object to a data.frame. Each row represents a scan, which can however be repeated several times if the PSM matches multiple proteins and/or contains two or more modifications. To reduce the data.frame so that rows/scans are unique and use semicolon-separated values to combine information pertaining a scan, use reduce.

Arguments

from
An object of class mzRident defined in the mzR package.

Details

See also the Tandem MS identification data section in the MSnbase-demo vignette.

Value

A data.frame

Author(s)

Laurent Gatto

Examples

## find path to a mzIdentML file
identFile <- dir(system.file(package = "MSnbase", dir = "extdata"),
                 full.name = TRUE, pattern = "dummyiTRAQ.mzid")
library("mzR")
x <- openIDfile(identFile)
x
as(x, "data.frame")

[Package MSnbase version 2.12.0 Index]