Navigate an

Navigate an MSnExp object by moving to the next or previous spectrum.

navMS(i, object, msLevel, nav = c("nextMS", "prevMS"), ...)

nextMS(...)

prevMS(...)

Arguments

i

The name or index of the current spectrum

object

The MSnExp object

msLevel

The MS level of the next or previous spectrum. If missing (default), then the level of the current spectrum is used.

nav

One of "nextMS" or "prevMS", to obtain the next or previous spectrum of level msLevel.

...

Additional parameters. Currently ignored.

Value

An object of class Spectrum1 or Spectrum2, depending on the value of msLevel or NULL, of no spectrum is found.

Examples

f <- msdata::proteomics(full.names = TRUE)[1]
x <- readMSData2(f, centroided = c(FALSE, TRUE, FALSE))

#> Error in data.frame(..., check.names = FALSE): arguments imply differing number of rows: 0, 2

sp <- "X009.1"
x[[sp]] ## curent MS3

#> Error in eval(expr, envir, enclos): object 'x' not found

MSnbase:::nextMS(sp, x) ## next MS3

#> Error in featureNames(object): object 'x' not found

MSnbase:::prevMS(sp, x) ## prev MS3

#> Error in featureNames(object): object 'x' not found

MSnbase:::prevMS(sp, x, 2L) ## prev MS2

#> Error in featureNames(object): object 'x' not found

MSnbase:::prevMS(sp, x, 1L) ## prev MS1

#> Error in featureNames(object): object 'x' not found