normToReference {MSnbase}	R Documentation

Combine peptides into proteins.

Description

This function combines peptides into their proteins by normalising the intensity values to a reference run/sample for each protein.

Usage

normToReference(x, group, reference = .referenceFractionValues(x = x, group =
  group))

Arguments

x
matrix, exprs matrix of an MSnSet object.

group
double or factor, grouping variable, i.e. protein accession; has to be of length equal nrow(x).

reference
double, vector of reference values, has to be of the same length as group and nrow(x).

Details

This function is not intented to be used directly (that's why it is not exported via NAMESPACE). Instead the user should use combineFeatures.

The algorithm is described in Nikolovski et al., briefly it works as follows:

Find reference run (column) for each protein (grouped rows). We use the run (column) with the lowest number of NA. If multiple candidates are available we use the one with the highest intensity. This step is skipped if the user use his own reference vector.

For each protein (grouped rows) and each run (column):

Find peptides (grouped rows) shared by the current run (column) and the reference run (column).

Sum the shared peptides (grouped rows) for the current run (column) and the reference run (column).

The ratio of the shared peptides (grouped rows) of the current run (column) and the reference run (column) is the new intensity for the current protein for the current run.

Value

a matrix with one row per protein.

Author(s)

Sebastian Gibb <mail@sebastiangibb.de>, Pavel Shliaha

References

Nikolovski N, Shliaha PV, Gatto L, Dupree P, Lilley KS. Label-free protein quantification for plant Golgi protein localization and abundance. Plant Physiol. 2014 Oct;166(2):1033-43. DOI: 10.1104/pp.114.245589. PubMed PMID: 25122472.

See Also

combineFeatures

Examples

library("MSnbase")
data(msnset)

# choose the reference run automatically
combineFeatures(msnset, groupBy=fData(msnset)$ProteinAccession)

# use a user-given reference
combineFeatures(msnset, groupBy=fData(msnset)$ProteinAccession,
 reference=rep(2, 55))

[Package MSnbase version 2.4.2 Index]