Reads as set of XML-based mass-spectrometry data files and generates an "MSnExp" object. This function uses the functionality provided by the mzR package to access data and meta data in mzData, mzXML and mzML.

readMSData(files, pdata = NULL, msLevel. = 2, verbose =
isMSnbaseVerbose(), centroided. = NA, smoothed. = NA, cache. = 1)

readMSData2(files, pdata = NULL, msLevel., verbose = isMSnbaseVerbose(),
centroided., smoothed. = NA)

Arguments

files

A character vector with file names to be read.

pdata

An object of class "NAnnotatedDataFrame" or NULL.

msLevel.

MS level spectra to be read. Use 1 for MS1 spectra or any larger numeric for MSn spectra. Default is 2. readMSData2 supports multiple levels and will, by default, read all the data.

centroided.

A logical in readMSData, indicating whether spectra are centroided or not. Default is NA. In readMSData2, the parameter value is missing and the value is set to NA in the returned object. It can be set for different MS levels by a vector of logicals, where the first element is for MS 1, the second element is for level 2, ... See "OnDiskMSnExp" for an example.

smoothed.

A logical indicating whether spectra already smoothed or not. Default is NA in readMSData and readMSData2.

cache.

Numeric indicating caching level. Default is 0 for MS1 and 1 MS2 (or higher). Under development.

verbose

Verbosity flag. Default is TRUE.

Details

For readMSData the whole MS data is read from file and kept in memory as Spectrum objects within the MSnExp's assayData slot.

To reduce the memory footpring especially for large MS 1 data sets it is also possible to read only selected information from the MS files and fetch the actual spectrum data (i.e. the M/Z and intensity values) only on demand from the original data files. This can be achieved with readMSData2. The function returns then an OnDiskMSnExp object instead of a MSnExp object.

Value

An MSnExp object for readMSData and a OnDiskMSnExp object for readMSData2.

See also

"MSnExp" or readMgfData to read mgf peak lists.

Examples

file <- dir(system.file(package = "MSnbase", dir = "extdata"),
            full.name = TRUE,
            pattern = "mzXML$")
aa <- readMSData(file)
aa

#> MSn experiment data ("MSnExp")
#> Object size in memory: 0.18 Mb
#> - - - Spectra data - - -
#>  MS level(s): 2 
#>  Number of spectra: 5 
#>  MSn retention times: 25:1 - 25:2 minutes
#> - - - Processing information - - -
#> Data loaded: Wed Feb  1 15:38:46 2017 
#>  MSnbase version: 2.1.10 
#> - - - Meta data  - - -
#> phenoData
#>   rowNames: dummyiTRAQ.mzXML
#>   varLabels: sampleNames
#>   varMetadata: labelDescription
#> Loaded from:
#>   dummyiTRAQ.mzXML 
#> protocolData: none
#> featureData
#>   featureNames: X1.1 X2.1 ... X5.1 (5 total)
#>   fvarLabels: spectrum
#>   fvarMetadata: labelDescription
#> experimentData: use 'experimentData(object)'

bb <- readMSData2(file)
bb

#> MSn experiment data ("OnDiskMSnExp")
#> Object size in memory: 0.03 Mb
#> - - - Spectra data - - -
#>  MS level(s): 2 
#>  Number of spectra: 5 
#>  MSn retention times: 25:1 - 25:2 minutes
#> - - - Processing information - - -
#> Data loaded [Wed Feb  1 15:38:47 2017] 
#>  MSnbase version: 2.1.10 
#> - - - Meta data  - - -
#> phenoData
#>   rowNames: dummyiTRAQ.mzXML
#>   varLabels: sampleNames
#>   varMetadata: labelDescription
#> Loaded from:
#>   dummyiTRAQ.mzXML 
#> protocolData: none
#> featureData
#>   featureNames: X1.1 X2.1 ... X5.1 (5 total)
#>   fvarLabels: fileIdx spIdx ... spectrum (26 total)
#>   fvarMetadata: labelDescription
#> experimentData: use 'experimentData(object)'