readMzTabData {MSnbase}	R Documentation

Read an 'mzTab' file

Description

This function can be used to create an "MSnSet" by reading and parsing an mzTab file. The metadata section is always used to populate the MSnSet's experimentData()@other$mzTab slot.

Usage

readMzTabData(file, what = c("PRT", "PEP", "PSM"), version = c("1.0",
  "0.9"), verbose = isMSnbaseVerbose())

Arguments

file
A character with the mzTab file to be read in.

what
One of "PRT", "PEP" or "PSM", defining which of protein, peptide PSMs section should be returned as an MSnSet.

version
A character defining the format specification version of the mzTab file. Default is "1.0". Version "0.9" is available of backwards compatibility. See readMzTabData_v0.9 for details.

verbose
Produce verbose output.

Value

An instance of class MSnSet.

Author(s)

Laurent Gatto

See Also

See MzTab and MSnSetList for details about the inners of readMzTabData.

Examples

testfile <- "https://raw.githubusercontent.com/HUPO-PSI/mzTab/master/examples/1_0-Proteomics-Release/PRIDE_Exp_Complete_Ac_16649.xml-mztab.txt"
prot <- readMzTabData(testfile, "PRT")
prot
head(fData(prot))
head(exprs(prot))
psms <- readMzTabData(testfile, "PSM")
psms
head(fData(psms))

[Package MSnbase version 2.12.0 Index]