readMzTabData_v0.9 {MSnbase}	R Documentation

Read an 'mzTab' file

Description

This function can be used to create a "MSnSet" by reading and parsing an mzTab file. The metadata section is always used to populate the MSnSet's experimentData slot.

Usage

readMzTabData_v0.9(file, what = c("PRT", "PEP"),
  verbose = isMSnbaseVerbose())

Arguments

file
A character with the mzTab file to be read in.

what
One of "PRT" or "PEP", defining which of protein of peptide section should be parse. The metadata section, when available, is always used to populate the experimentData slot.

verbose
Produce verbose output.

Value

An instance of class MSnSet.

Author(s)

Laurent Gatto

See Also

writeMzTabData to save an "MSnSet" as an mzTab file.

Examples

testfile <- "https://raw.githubusercontent.com/HUPO-PSI/mzTab/master/legacy/jmztab-1.0/examples/mztab_itraq_example.txt"
prot <- readMzTabData_v0.9(testfile, "PRT")
prot
pep <- readMzTabData_v0.9(testfile, "PEP")
pep

[Package MSnbase version 2.12.0 Index]