Reduce a data.frame

Reduce a data.frame so that the (primary) key column contains only unique entries and other columns pertaining to that entry are combined into semicolon-separated values into a single row/observation.

# S4 method for data.frame
reduce(x, key, sep = ";")

Arguments

x
A data.frame.

key
The column name (currenly only one is supported) to be used as primary key.

sep
The separator. Default is ;.

Value

A reduced data.frame.

Examples

dfr <- data.frame(A = c(1, 1, 2),
                  B = c("x", "y", "z"),
                  C = LETTERS[1:3])
dfr

#>   A B C
#> 1 1 x A
#> 2 1 y B
#> 3 2 z C

reduce(dfr, key = "A")

#> Error in reduce(dfr, key = "A"): could not find function "reduce"