Removes non-identified features

The method removes non-identifed features in MSnExp and MSnSet instances using relevant information from the feaureData slot of a user-provide filtering vector of logicals.

Methods

signature(object = "MSnExp", fcol = "pepseq", keep = NULL)

Removes the feature from object that have a feature fcol (default is "pepseq") equal to NA. Alternatively, one can also manually define keep, a vector of logical, defining the feature to be retained.

signature(object = "MSnSet", fcol = "pepseq", keep = NULL)

As above of MSnSet instances.

See also

MSnExp and MSnSet.

Examples

  quantFile <- dir(system.file(package = "MSnbase", dir = "extdata"),
                     full.name = TRUE, pattern = "mzXML$")
  identFile <- dir(system.file(package = "MSnbase", dir = "extdata"),
                     full.name = TRUE, pattern = "dummyiTRAQ.mzid")
  msexp <- readMSData(quantFile)
  msexp <- addIdentificationData(msexp, identFile)
  fData(msexp)$sequence

#> [1] "VESITARHGEVLQLRPK" "IDGQWVTHQWLKK"     NA                 
#> [4] NA                  "LVILLFR"          

  length(msexp)

#> [1] 5

  ## using default fcol
  msexp2 <- removeNoId(msexp)
  length(msexp2)

#> [1] 3

  fData(msexp2)$sequence

#> [1] "VESITARHGEVLQLRPK" "IDGQWVTHQWLKK"     "LVILLFR"          

  ## using keep
  print(fvarLabels(msexp))

#>  [1] "spectrum"                 "sequence"                
#>  [3] "chargeState"              "rank"                    
#>  [5] "passThreshold"            "experimentalMassToCharge"
#>  [7] "calculatedMassToCharge"   "modNum"                  
#>  [9] "isDecoy"                  "post"                    
#> [11] "pre"                      "start"                   
#> [13] "end"                      "DatabaseAccess"          
#> [15] "DBseqLength"              "DatabaseSeq"             
#> [17] "DatabaseDescription"      "idFile"                  
#> [19] "MS.GF.RawScore"           "MS.GF.DeNovoScore"       
#> [21] "MS.GF.SpecEValue"         "MS.GF.EValue"            
#> [23] "modName"                  "modMass"                 
#> [25] "modLocation"              "subOriginalResidue"      
#> [27] "subReplacementResidue"    "subLocation"             
#> [29] "nprot"                    "npep.prot"               
#> [31] "npsm.prot"                "npsm.pep"                

  (k <- fData(msexp)$'MS.GF.EValue' > 75)

#> [1]  TRUE FALSE    NA    NA  TRUE

  k[is.na(k)] <- FALSE
  k

#> [1]  TRUE FALSE FALSE FALSE  TRUE

  msexp3 <- removeNoId(msexp, keep = k)
  length(msexp3)

#> [1] 2

  fData(msexp3)$sequence

#> [1] "VESITARHGEVLQLRPK" "LVILLFR"