Removes non-identified features

The method removes non-identifed features in MSnExp and MSnSet instances using relevant information from the feaureData slot of a user-provide filtering vector of logicals.

Methods

signature(object = "MSnExp", fcol = "pepseq", keep = NULL)

Removes the feature from object that have a feature fcol (default is "pepseq") equal to NA. Alternatively, one can also manually define keep, a vector of logical, defining the feature to be retained.

signature(object = "MSnSet", fcol = "pepseq", keep = NULL)

As above of MSnSet instances.

See also

MSnExp and MSnSet.

Examples

  quantFile <- dir(system.file(package = "MSnbase", dir = "extdata"),
                     full.name = TRUE, pattern = "mzXML$")
  identFile <- dir(system.file(package = "MSnbase", dir = "extdata"),
                     full.name = TRUE, pattern = "dummyiTRAQ.mzid")
  msexp <- readMSData(quantFile)

#> Reading 5 MS2 spectra from file dummyiTRAQ.mzXML
#>   |                                                                              |                                                                      |   0%  |                                                                              |==============                                                        |  20%  |                                                                              |============================                                          |  40%  |                                                                              |==========================================                            |  60%  |                                                                              |========================================================              |  80%  |                                                                              |======================================================================| 100%
#> Creating 'MSnExp' object

  msexp <- addIdentificationData(msexp, identFile)

#> reading dummyiTRAQ.mzid... DONE!

  fData(msexp)$pepseq

#> [1] "VESITARHGEVLQLRPK" "IDGQWVTHQWLKK"     NA                 
#> [4] NA                  "LVILLFR"          

  length(msexp)

#> [1] 5

  ## using default fcol
  msexp2 <- removeNoId(msexp)
  length(msexp2)

#> [1] 3

  fData(msexp2)$pepseq

#> [1] "VESITARHGEVLQLRPK" "IDGQWVTHQWLKK"     "LVILLFR"          

  ## using keep
  print(fvarLabels(msexp))

#>  [1] "spectrum"                  "scan number(s)"           
#>  [3] "passthreshold"             "rank"                     
#>  [5] "calculatedmasstocharge"    "experimentalmasstocharge" 
#>  [7] "chargestate"               "ms-gf:denovoscore"        
#>  [9] "ms-gf:evalue"              "ms-gf:rawscore"           
#> [11] "ms-gf:specevalue"          "assumeddissociationmethod"
#> [13] "isotopeerror"              "isdecoy"                  
#> [15] "post"                      "pre"                      
#> [17] "end"                       "start"                    
#> [19] "accession"                 "length"                   
#> [21] "description"               "pepseq"                   
#> [23] "modified"                  "modification"             
#> [25] "idFile"                    "databaseFile"             
#> [27] "nprot"                     "npep.prot"                
#> [29] "npsm.prot"                 "npsm.pep"                 

  (k <- fData(msexp)$'ms-gf:evalue' > 75)

#> [1]  TRUE FALSE    NA    NA  TRUE

  k[is.na(k)] <- FALSE
  k

#> [1]  TRUE FALSE FALSE FALSE  TRUE

  msexp3 <- removeNoId(msexp, keep = k)
  length(msexp3)

#> [1] 2

  fData(msexp3)$pepseq

#> [1] "VESITARHGEVLQLRPK" "LVILLFR"