selectFeatureData {MSnbase}	R Documentation

Select feature variables of interest

Description

Select feature variables to be retained.

Usage

selectFeatureData(object, graphics = TRUE, fcol)

Arguments

object
An MSnSet.

graphics
A logical (default is TRUE) indicating whether a shiny application should be used if available. Otherwise, a text menu is used. Ignored if k is not missing.

fcol
A numeric, local or character of valid feature variables to be passed directly.

Value

Updated MSnSet, containing only selected feature variables.

Author(s)

Laurent Gatto

Examples

library("pRolocdata")
data(hyperLOPIT2015)
## 5 first feature variables
x <- selectFeatureData(hyperLOPIT2015, fcol = 1:5)
fvarLabels(x)
## Not run: 
## select via GUI
x <- selectFeatureData(hyperLOPIT2015)
fvarLabels(x)

## End(Not run)

[Package MSnbase version 2.4.2 Index]